chemical-analysis-project/backend/requirements.txt

flask
flask-sqlalchemy
rdkit-pypi
flask-cors
